lfs,tq: move type *lfs.TransferQueue to pkg "tq"
