+++ lfs/download_queue.go
@@ -2,6 +2,7 @@ package lfs
 
 import (
 	"github.com/git-lfs/git-lfs/api"
+	"github.com/git-lfs/git-lfs/tq"
 	"github.com/git-lfs/git-lfs/transfer"
 )
 
@@ -40,11 +41,11 @@ func NewDownloadable(p *WrappedPointer) *Downloadable {
 }
 
 // NewDownloadCheckQueue builds a checking queue, checks that objects are there but doesn't download
-func NewDownloadCheckQueue(options ...transferQueueOption) *TransferQueue {
-	return newTransferQueue(transfer.Download, options...)
+func NewDownloadCheckQueue(options ...tq.TransferQueueOption) *tq.TransferQueue {
+	return tq.NewTransferQueue(transfer.Download, options...)
 }
 
 // NewDownloadQueue builds a DownloadQueue, allowing concurrent downloads.
-func NewDownloadQueue(options ...transferQueueOption) *TransferQueue {
-	return newTransferQueue(transfer.Download, options...)
+func NewDownloadQueue(options ...tq.TransferQueueOption) *tq.TransferQueue {
+	return tq.NewTransferQueue(transfer.Download, options...)
 }
