+++ lfs/upload_queue.go
@@ -8,6 +8,7 @@ import (
 	"github.com/git-lfs/git-lfs/api"
 	"github.com/git-lfs/git-lfs/config"
 	"github.com/git-lfs/git-lfs/errors"
+	"github.com/git-lfs/git-lfs/tq"
 	"github.com/git-lfs/git-lfs/transfer"
 )
 
@@ -67,8 +68,8 @@ func NewUploadable(oid, filename string) (*Uploadable, error) {
 }
 
 // NewUploadQueue builds an UploadQueue, allowing `workers` concurrent uploads.
-func NewUploadQueue(options ...transferQueueOption) *TransferQueue {
-	return newTransferQueue(transfer.Upload, options...)
+func NewUploadQueue(options ...tq.TransferQueueOption) *tq.TransferQueue {
+	return tq.NewTransferQueue(transfer.Upload, options...)
 }
 
 // ensureFile makes sure that the cleanPath exists before pushing it.  If it
