+++ tq/transfer_queue_test.go
@@ -1,4 +1,4 @@
-package lfs
+package tq
 
 import (
 	"testing"
